requirements.txt
attrs==19.1.0
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django
django-filter==2.0.0
django-filters==0.2.1
django-import-export==1.2.0
django-jet==1.0.8
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sendsms==0.3.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0
et-xmlfile==1.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4
jet==1.0
Jinja2==2.10.1
jsonschema==3.0.1
lml==0.0.9
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.14
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-ods==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
PyJWT==1.7.1
pyrsistent==0.14.11
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel.yaml==0.15.91
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
tablib==0.13.0
texttable==1.6.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
